+++ src/types/preferences.tsx
@@ -33,5 +33,6 @@ export interface Preferences {
   useBrowserExtension: boolean;
   searchRecursively: boolean;
   checkDuplicates: boolean;
+  smartTitles: boolean;
   faviconField?: string;
 }